Extract protocols as openai-protocol workspace crate
Move protocol definitions from src/protocols to a standalone workspace
crate named openai-protocol. This enables:

- Independent versioning and publishing to crates.io
- Reuse by other projects needing OpenAI-compatible types
- Cleaner dependency graph for the main crate

Changes:
- Create protocols/ workspace member with its own Cargo.toml
- Move all protocol types (chat, completion, embedding, etc.)
- Add axum as optional feature for ValidatedJson extractor
- Move IntoResponse impls for FlushCacheResult/WorkerLoadsResult
- Move JobStatus constructor methods to protocols crate
- Re-export as `pub use openai_protocol as protocols` for compatibility
- Exclude bindings/python from workspace to fix maturin builds
