+++ protocols/Cargo.toml
@@ -1,0 +1,23 @@
+[package]
+name = "openai-protocol"
+version = "0.1.0"
+edition = "2021"
+description = "OpenAI-compatible API protocol definitions and types"
+license = "Apache-2.0"
+repository = "https://github.com/lightseekorg/smg"
+keywords = ["openai", "llm", "api", "protocol", "types"]
+categories = ["api-bindings", "data-structures"]
+
+[features]
+default = []
+axum = ["dep:axum"]
+
+[dependencies]
+serde = { version = "1.0", features = ["derive"] }
+serde_json = { version = "1.0", features = ["preserve_order"] }
+validator = { version = "0.20.0", features = ["derive"] }
+axum = { version = "0.8.6", optional = true }
+rand = "0.9.2"
+tracing = "0.1"
+chrono = "0.4"
+tokio = "1.49.0"
