+++ protocols/src/chat.rs
@@ -13,7 +13,7 @@ use super::{
     },
     sampling_params::{validate_top_k_value, validate_top_p_value},
 };
-use crate::protocols::{
+use crate::{
     builders::{ChatCompletionResponseBuilder, ChatCompletionStreamResponseBuilder},
     validated::Normalizable,
 };
