+++ protocols/src/common.rs
