+++ protocols/src/embedding.rs
