+++ src/lib.rs
@@ -10,7 +10,7 @@ pub mod middleware;
 pub mod multimodal;
 pub mod observability;
 pub mod policies;
-pub mod protocols;
+pub use openai_protocol as protocols;
 pub mod reasoning_parser;
 pub mod routers;
 pub mod server;
